{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "0.1.0"
description = ""
authors = ["ali.rasoulinejad <ali.rasoulinejad75@gmail.com>"]
readme = "README.md"
packages = [{include = "django_template"}]

[tool.poetry.dependencies]
python = "^3.10"
django = ">=4.2,<5.0"
djangorestframework = ">=3.14,<4.0"
django-redis = ">=5.2,<6.0"
django-environ = ">=0.10,<1.0"
drf-spectacular = ">=0.26,<1.0"
psycopg2-binary = ">=2.9,<3.0"
mypy = ">=1.3,<2.0"
django-stubs = "^4.2.1"
djangorestframework-stubs = "^3.14.1"
result = "^0.10.0"
django-prometheus = "^2.3.1"
opentelemetry-instrumentation-django = "^0.39b0"
opentelemetry-instrumentation-psycopg2 = "^0.39b0"
opentelemetry-sdk = "^1.18.0"
opentelemetry-api = "^1.18.0"
opentelemetry-instrumentation = "^0.39b0"
opentelemetry-instrumentation-logging = "^0.39b0"
opentelemetry-exporter-jaeger = "^1.18.0"
flake8 = "^6.0.0"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.local.dependencies]
django-extensions = ">=3.2,<4.0"
django-debug-toolbar = ">=4.1,<5.0"


[tool.poetry.group.production.dependencies]
sentry-sdk = ">=1.25,<2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
